+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 - Update gemspec
+- Read certificate full path
+- Return early if get access token fail
 ### 2020-12-09
 - Reset configuration before new configuration
 - Encrypt initiator password with X506 Certificate key